refactor(eigendb): moving api endpoints to /api/v1 to start versioning the api
