+++ apps/eigendb/api/v1/embeddings/delete.go
