+++ apps/eigendb/api/v1/embeddings/insert.go
