+++ apps/eigendb/api/v1/embeddings/retrieve.go
