+++ apps/eigendb/api/v1/embeddings/upsert.go
