+++ apps/eigendb/api/v1/indexes/delete.go
