+++ apps/eigendb/api/v1/indexes/list.go
