+++ apps/eigendb/api/v1/indexes/stats.go
