+++ apps/eigendb/e2e/health_testsuite.venom.yml
@@ -1,7 +1,7 @@
 name: Health Test Suite
 
 vars:
-  url: http://localhost:8080
+  url: http://localhost:8080/api/v1
 
 testcases:
 - name: Test health check
