+++ apps/eigendb/e2e/indexes_testsuite.venom.yml
@@ -1,7 +1,7 @@
 name: Indexes Test Suite
 
 vars:
-  url: http://localhost:8080
+  url: http://localhost:8080/api/v1
 
 testcases:
 - name: Test create index
