+++ apps/eigendb/e2e/middleware_testsuite.venom.yml
@@ -1,7 +1,7 @@
 name: API Middleware Test Suite
 
 vars:
-  url: http://localhost:8080
+  url: http://localhost:8080/api/v1
 
 testcases:
 - name: Authenticated request using valid API key
